p4testgen: batch 155 tests into a single JVM
Each p4_testgen_test() created its own kt_jvm_test target and JVM,
so `bazel test //...` spawned 155+ JVMs in parallel — exhausting memory.

Add a p4_testgen_suite() macro that creates the per-program build rules
(p4testgen + p4c compilation, still parallelized) but collects all outputs
into a single kt_jvm_test. A new P4TestgenSuiteTest discovers all *_stfs/
directories at runtime and parameterizes as "program/stfName".

Also raise detekt NestedBlockDepth threshold to 5 (buildTraceTree's
while → when → try/catch is naturally 5 levels deep).
